+++ src/test/java/seedu/duke/DukeTest.java
@@ -1,4 +1,4 @@
-package BookBob;
+package seedu.duke;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
